+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to the "Project Timer" extension will be documented in this 
     - Before: `0` would disable idle detection.
     - After: `0` will be treated as `0` minutes (pause immediately when idle).
 - Change project name style on status bar menu.
+- Change `project-timer.statusBar.displayToday` into `project-timer.statusBar.displayTimeMode` to support more optional display modes.
 
 ## [0.0.1] - 2026-02-12
 Initial release of Project Timer.